# flow2print-platform/schemas/openapi/edge-api.yaml
openapi: 3.1.0
info:
  title: Flow2Print Edge API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /v1/projects/{id}/artifacts:
    get:
      summary: List artifacts for a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artifact list returned
  /v1/projects/{id}/preflight:
    get:
      summary: Get latest preflight report for a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Preflight report returned
...
  /v1/connectors/magento2/projects/{projectId}/status:
    get:
      summary: Get Magento-facing status for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project status returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  projectVersionId:
                    type: string
                  status:
                    type: string
                  state:
                    type: string
                  approvalState:
                    type: string
                  preflightStatus:
                    type: string